Authentication and authorization using database users
The API does HTTP Basic auth and uses the credentials in each request to
authenticate with the database.  This authenticated session is then used
to perform actions and the database role system is responsible for
authorization.

While it may be a little unusual these days for an application to rely
on database users and roles instead of re-creating those concepts
itself, there are many benefits to this approach for our use case.  In
particular, it's an expedient way to handle authn/authz without lots of
additional code, and it allows us to very easily implement a standard
role-based access control (RBAC) system leveraging Pg's fine-grained
permission system.  The downsides, for example not being able to easily
re-use connections, are minimal for our low-traffic use case; the
overhead of a new connection each time is basically equivalent to what
we'd incur if our API was in turn talking HTTP to something else, a
common and accepted pattern.