Manage schema with Sqitch
Sqitch <https://sqitch.org> is a tool I've used a lot in the past to
manage database changes, and it really shines.  I don't know why I
didn't think to start prototyping with it since I knew I wanted to use
it eventually.  It's easiest to start with it from the beginning, so
this ditches the little bin/create-database script and sql.d/ files and
moves to Sqitch.

Roles are re-designed in this commit to more closely match how I've
designed the API access controls to work.