Beginnings of a "warehouse" schema
This schema is used to hold the warehoused relational data after it's
gone through our ETL-processes from the receiving schema.

I came to the design here after sketching out the various options and
trying them out.  More to come, of course, but first I'm going to switch
to sketching out the basic ETL process.  Then both the schema and ETL
process can be developed in tandem.