Rename "staging" schema to "receiving"
While ETL often refers to "staged" data, it seems to me that terms from
distribution centers (DC) make more sense for what's happening in this
database.  Data is received, goes through an ETL process to be
warehoused, and then is provided ("shipped") back out for downstream
research and analysis.  I also want to avoid confusion with the term
"staging" as used for pre-production deployment systems, and think the
DC terms are more obvious.

My plan is to have these three schemas (and maybe more):

  • receiving — incoming data, minimally- or un-processed
  • warehouse — structured relational data with strong consistency
                and semantics
  • shipping — tables and views to support outgoing data needs

with ETL processes between them as necessary.

Deployed databases will need to be:

  1. sqitch reverted all the way back to the beginning using the
     previous commit
  2. deployed all the way forward from this commit

This is only ok because we're in very early phases still and there is no
production data to consider.