receiving.enrollment: Switch from a single "processed" timestamp to a JSON "processing_log"
This gives us space to process enrollments more than once as our ETL
routine changes.  This column is used as a marker of ETL progress, and
switching to a log gives us an array of markers.

The log is kept as a JSON column instead of a separate table to allow
arbitrary metadata to be stored by the ETL routine, and also to avoid
setting a precedent of having an additional log table per receiving
table (or using a single log table and bypassing relational integrity).