etl enrollments: Allow site type to be missing
Even though "type" should be guaranteed by the Encounter protocol
defined by the Audere app, we received a document yesterday that is
missing "type".  Type-checking on the app side must be violated/ignored
at some point!  It is not crucial that we have this, so be a little more
forgiving.