Log all messages to the local syslog service if available
It's super-useful for automated processes, especially those using cron,
to be quiet on the console but chatty in system logs.  Such behaviour
aids debugging when unexpected problems crop up.  The LOG_LEVEL
environment variable can be set to "warning" or "error" for cron jobs to
avoid email notifications from cron when everything goes swimmingly.
When things blow up, the gory details will be in syslog.