api.datastore: Use db.session instead of connecting directly
db.session started as a copy of the datastore module and has been
refined a bit since then for the ETL commands.  Now that the API and
database code is in the same repo, it makes sense to use it for the
api.datastore functions too.  In particular, this gives us access to
savepoints which I'll be using shortly for identifier minting.