Receive presence/absence calls
These calls will come from Samplify (an in-house LIMS), which makes them
for our samples based on a TaqMan qPCR panel run at the Northwest
Genomics Center.  The call document and API request data is not
validated other than the database ensuring it is JSON.  Similar to
enrollments, we'll use an ETL routine to copy the relevant data into.

I initially wanted to define a pair of endpoints, one taking
well-structured, JSON Schema-validated presence/absence calls and the
other raw qPCR results as RDML straight from the instrument, but the
Samplify team balked.  They saw this as additional work that wasn't
useful to them.  Instead, we'll be receiving their bespoke JSON
representation of the information RDML already models.  I still think
that this would be a good approach in the future and set us up for
integrating with other diagnostic labs.