Receive sequence read set notifications
Notification of available sequence read data will come from Samplify (an
in-house LIMS).  The read set document and API request data is not
validated other than the database ensuring it is JSON.  Similar to
enrollments, we'll use an ETL routine to trigger relevant data into the
warehouse.

The request data this endpoint accepts is intentionally very open-ended
right now.  We're still in an exploration and discovery phase about what
will work best, and we can add validation later as we settle on
conventions.  For example, while it's clear we want the read set to
contain a list of URLs pointing at FASTQs we can fetch, it's not clear
how the other metadata should be structured yet.  Part of the philosophy
of the receiving area is that it is very liberal in what it accepts on
the premise that we want to collect the data first and then decide how
to handle it.