Require a non-empty username/password for API auth
To my surprise, the underlying database driver treats an empty string
password the same as a missing password (None).  When the driver
receives a username with no password, it looks for and uses matching
credentials provided in an optional pgpass file.  We don't want to allow
system credentials stored in any pgpass file to be available for use by
an API client, so a password is always required.