Add ISO week of encounter to incidence model observation view
Time binning is useful for the modeling and viz efforts, so it's nice to
harmonize that early in the data flow rather than separately several
times.  Weeks are the standard time bin used in epidemiology.

Note that the WHO's epi weeks, synonymous with ISO weeks, are slightly
different from CDC's epi weeks, as described most succinctly¹

> The US CDC defines epidemiological week (known as MMWR week) as seven
> days beginning with Sunday and ending with Saturday. The WHO defines
> epidemiological week, based on ISO week, as seven days beginning with
> Monday and ending with Sunday. In either case, the end of the first
> epidemiological week of the year by definition must fall at least four
> days into the year. Most years have 52 epidemiological weeks, but some
> have 53.

It's my understanding that WHO / ISO weeks are used more widely around
the world, and just like we use the metric system in science, it seems
we should use international datetime standards as well.

If we find ourselves wanting CDC epi weeks in the future, we can add
them as well.

¹ https://epiweeks.readthedocs.io