Add schema and ETL to store presence/absence data in warehouse
Three tables were added to the warehouse schema:
* sample
* target
* presence_absence

These tables are designed to be in the simplest form to store relevant
presence_absence data received from Samplify. Future needs will lead to
modifications.

The ETL process performs a `find_or_create` for targets, as we
anticipate the same targets to be tested, but leave room for the
addition of new targets.

The ETL process performs a `find_or_create` for samples to allow the
processing of re-tested samples. We do not expect to update information
of a sample itself.

The ETL process performs an `upsert` for presence_absence to allow the
update of results should there be a re-test on an old sample.

The presence absence ETL process will abort under these conditions:
1. If a sample's barcode matches with a UUID that is of the incorrect identifier set
2. If we receive an unexpected value for the "controlStatus" of a target
3. If we receive an unexpected value for the "targetResult" of a specific test