Remove unused /scan API endpoint and associated receiving tables
Direct entry of barcode scans, which this code was meant to support,
never panned out as we embarked on the study.  It is still desired by
our team and the lab groups, but the requirement for collecting other
associated data at the same time made the scope too large for this year.

Instead, manifests of scans were kept in Excel workbooks on OneDrive
(alongside PHI/PII), and the infrastructure in ID3C for processing those
is too follow.

This reverts the receiving/scan sqitch change by reworking it and
swapping the deploy and revert scripts.