etl presence-absence: Require samples to already exist
Samples will no longer be created by this ETL routine, as the plan is
for samples to come in only via enrollments and sample manifests.  This
routine will abort if a sample that ostensibly should exist is not
found.  This isn't expected to happen very often, but it could if a
manifest load hasn't happened for a sample that's already run through
processing.  As soon as the sample exists, this routine will be able to
proceed.  We can improve the restart/retry process in the future.

Existing samples will have their details updated to include the
information we're grabbing from presence/absence results.

As a consequence of this change, results with an unknown/invalid sample
identifier (distinct from results with a known/valid identifier but
without a sample record) are now skipped with a warning.  Previously
samples were created for them using the unknown/invalid barcode.  This
should be ok, and will naturally filter out ad-hoc quality control and
assay development samples.

If this routine was allowed to create samples, we could end up with two
sample records (one with only a sample identifier, one with only a
collection identifer) that would ultimately need complicated merging
during an ETL routine.  There are other solutions if our needs require
them, but this is the simplest one I could come up with for now.

Revision bump means this will re-process all results.  Before being run,
it would be best to truncate the warehouse.sample table to avoid a
merging problem.