Refactor grants to enrollment-processor into a reworkable sqitch change
This makes it easier to update the grants for this role and see all of
them in one place.  I think it will be a better pattern going forward
than what I've done previously, which is add changes like
roles/incidence-modeler/grant-select-on-age-bins.

I'll be editing this change in-place in subsequent commits on this
branch.

For future roles, I plan to split the initial sqitch change into a
creation change and a grants change, since creation is not easy to make
idempotent.