Handle the flu@home response token ("dontKnow") indicating an unknown value
This token is used by the flu@home questionnaires and is distinct from
the "doNotKnow" token used by kiosk questionnaires.  🙄  Problems like
this will go away once we model the questionnaire data relationally and
don't have to deal with raw values.

I ran into it while testing this branch locally, and fixed it here
because the error prevents the view from working.