etl enrollments: Skip processing samples without a barcode
A record but no barcode means that a sample was collected, but we don't
know for sure what the barcode was because of some mixup or error.
Until corrections applied this morning, this kind of record didn't exist
(and the protocol spec doesn't allow for it, but we already know that
spec isn't enforced).