Remove clinical data instances without barcode
Added parameter when reading data from UW so that empty strings and 'NA'
evaluate to null values.

Dropped data with missing barcodes from the pipeline so they don't get
uploaded to the database.

Added an option to the preprocess CLI to allow the user to provide an
output filename for data with missing barcodes. If the user provides no
option for output, the CSV of missing barcodes gets dumpted to stdout.

Additionally, refactored the main method in the clinical preprocess
step to be shorter and easier to read.

Co-author: Jover Lee <joverlee521@gmail.com>

Drop duplicate barcodes in clinical data

In UW-clinical preprocessing, keep only the first barcode when there are
multiple barcodes matching an MRN + Accession number.

Add functionality for printing problematic barcodes in combination with
missing barocdes.

TODO: debug Pandas warning for slicing and editing a DataFrame.

Co-author: Jover Lee <joverlee521@gmail.com>

Address Pandas warnings