Remove 'Unknown'/'NULL' mapping in clinical etl
Add a `na_values` argument to the pandas read for importing UW clinical
records. 'NA'-type values are now handled in the clinical preprocess
when reading in the data using the pandas function `read_csv()`.

The etl clinical pipeline no longer needs to convert the specified
`na_values` (e.g. 'Unknown') to null.

Note: 'NULL' does not need to be added to `na_values`, because by
default, pandas converts 'NULL' to `null` if `keep_default_na` is kept
to `true`.