Convert dtypes post column renaming where possible
We are able to convert age (UW) and encounter date columns (UW and SCH)
to their appropriate dtypes post-renaming. This will reduce the number
of changes necessary in the event that clinical records column names are
changed again.

However, one area where we have to convert the dtype before renaming is
in the UW manifest data load. For some reason, pandas reads in UW
clinical record dates `object` and UW manifest dates as `dateTime`. So,
the `parse_date` argument is provided in pandas `read_csv` and
`read_excel` to correct this.