Add security definer to barcode trigger function
The `identifier-minter` role cannot run `identifier mint` due to a SQL
lock on table `warehouse.identifier`. This insufficient privilege error
arises when the `warehouse.identifier_barcode_distance_check()` function
is invoked.

Mark the function with `security definer` so the function is run as the
owner, not the user, to work around this.

Additionally, add a `truncate` statement to the barcode-distance-check
verify script. The test data in that file began conflicting with actual
data in the production database. Until we can guarantee that test
barcodes never conflict with real barcodes, the simplest way forward is
to temporarily delete data in the identifier table but rollback after
testing. This potentially dangerous operation is noted in comments in
the file.