Change all SET statements to SET LOCAL in schema change scripts
This isolates the effects of SET to each change's transaction for safety
and least surprise.  I'm not sure if each sqitch change is run in a
separate session or not, but regardless, it's good to avoid
action-at-a-distance.

Change scripts are modified in place and will only effect development
environments when they're redeployed.  New change scripts should use SET
LOCAL statements from the beginning.