Update modeling view v1
Update shipping.incidence_model_observation_v1
to use new age column in encounter table.

Currently rounds age to the nearest year and
converts to integer so that the data type does not
change for the modeling team.

Will add version 2 of the modeling view to use
intervals and intervalranges to account for
age in months for infants.