Add model view v2
Create incidence_model_observation_v2 to use
the new age column and use new age bins.

The major change in this version is that age is now
displayed as a float to allow the splitting of infants
by age in months.

The age ranges are displayed as intervals, but the
upper and lower bounds are displayed as floats
as well.