Stop using an automatic union merge for the Sqitch plan
While this is often desirable behaviour on merge to avoid a conflict, it
interacts poorly with rebasing of a branch when the plan is changed
during the rebase.  The result is often duplicated lines in the plan.
This may produce more merge conflicts, but at least a human will be
involved in sorting out the problem instead of the computer silently
doing it wrong.