Add Kit table to warehouse schema
Creates `kit` table in the warehouse to hold 
information about self-test kits from Flu@home.

Kits act as the connector between encounter and
the two samples collected via Flu@home kits. Each kit
references an encounter and up to 2 samples.

The idea is that a kit can be created via 
`etl enrollment` or `etl manifest`. If a kit is created
via `enrollment` then it will reference an encounter
but no samples. If a kit is created via `manifest` then
it will references samples but no encounter. The 
other etl process will fill in the missing references.