clinical parse-uw: Account for column name change in HMC manifest
This was accounted for in our separate manifest parsing, but not in the
different parsing of the manifest here.  The effect was that ~2,600 records
were skipped due to missing barcodes when reloading the clinical data
earlier today.

In the (hopefully near) future, we won't need to do the merge here and
this code will be deleted.