Add an organism table to the warehouse
The organism table represents a hierarchy of taxonomic lineages, such as
viral pathogens.  It will allow us to annotate presence/absence targets,
genomes (TK), and vaccinations (TK) with the appropriate lineage.  This,
in turn, will let us query for all Influenza A targets without needing
to know the particular target names used or all RSV genomes.

Organisms also provide a place to store ad-hoc details about a pathogen,
such as if it is considered reportable or not.

The label tree (ltree) data type is used to encode the hierarchy and act
as a unique key for each organism.  It allows fast, indexed queries on
the hierarchy itself with obvious semantics and without requiring
tedious and iterative chains of self-joins.  One downside is that it
does essentially denormalize the hierarchy, which could make changes to
the hierarchy a little more tedious.  This seems acceptable since the
hierarchies we expect to care about are pretty stable.

Organisms can (and generally should) refer out to other external
databases, such as NCBI Taxonomy (NCBITAXON)¹, using the identifiers
column.

¹ https://www.ncbi.nlm.nih.gov/taxonomy
  https://bioportal.bioontology.org/ontologies/NCBITAXON