manifest: Include self-test columns
Update both `manifest parse` and
`manifest parse-using-config` to include columns
related to the self-test manifest.

Note that the self-test manifest needs to be parsed
twice, once for "utm" samples and another time
for "rdt" samples. Remember to add the
`sample-type` to differentiate between the two 
types of samples!

"rdt" samples do not have a sample barcode, so use
the collection barcode as both the sample barcode and
the collection barcode in the config file.
 
`qc_barcodes` drops rows
with duplicated barcodes.

Will update `OneDrive.yaml` to include the new
columns for the self-test manifest.