Add etl kit
Add `kit` as a group within `etl` and add two sub-commands
`enrollments` and `manifest`.

The kit etl commands are completely dependent on the core
etl routines `etl manifest` and `etl enrollments` because
the kit etl commands do not create encounters or samples.
It only looks up existing encounters/samples within the
relational warehouse and creates kits with the appropriate
links to encounters/samples.

If the encounter/sample cannot be found within the warehouse,
the kit etl process will stop completely. The idea is that
the encounter/sample will eventually get created by the core
etl routines and the kit etl process can pick up where it
left off.

Both commands update samples linked to kits to link them to
the appropriate encounter. If the sample is already linked
to a different encounter, the etl process will error out.