Add genome to warehouse
A genome is an explicit representation of a whole 
genome of an organism that was extracted from
a sample. The genome may also reference a 
sequence read set if the genome was generated
from known sequence reads (e.g. consensus genomes). 

The `sequence_read_set_id` can be null to allow 
the inclusion of genomes from external sources that
do not specify how the genome was generated. 

Each genome must have a unique combination
of sample, organism, and sequence read set so
that the table does not have any duplicate genomes.