Incidence model views: Correct handling of encounters with multiple (or no) samples
The lateral subquery had two flaws which only became apparent with
Jover's self-test kit work that greatly increases the number of
encounters with multiple samples (RDT and UTM).

 1. A misunderstanding of lateral subqueries meant that encounters
    without a sample were being omitted because the join was inner
    not left.

 2. A bug in the use of the first_value() window function meant that
    encounters with multiple samples produced wholly duplicate rows in
    the view (one row per sample).  The intent, as noted in my original
    commit message, was to return a _single_ row for each encounter,
    with an arbitrarily-picked "first" sample.

Since the modeling team was already having to deal with multiple rows
for some encounters, correct the flawed queries in both v1 and v2 views
to report the left join of encounters with samples.  This adds in the
previously missing encounters and reports the actual sample identifier
for the previously duplicated encounters.