Add genomic sequence to warehouse
Genomic sequences reference the genome
that they belong to. To ensure that sequences
aren't repeats, their identifier has to be unique, 
and there can only be one sequence per segment
per genome. 

There was thought to add a separate table
for genome segments to standardize the 
nomenclature for segments, but as of right now
that seems to be over-complicating the database.
IF the need arises, the `segment` column can be
altered to `segment_id` to reference another 
table.