Add upload-consensus-genome
Uploads consensus genomes and assembly summaries
into the database receiving area. The consensus genomes
and summaries should be in a newline-delimited JSON 
format that matches the JSON payload created
during the assembly pipeline. 

This command is to be used for backfilling the
pre-existing genomic data and if we ever need 
to do a batch upload of consensus genomes.
Otherwise, the consensus genomes should be
uploaded to the receiving area via POST request
at the end of the assembly pipeline.