Add sequence read set table to warehouse
A sequence read set represents an entire set of
FASTQ files for a sample from one sequencing run.

The actual FASTQ files will live in a remote object
store and and the sequence read set will point to
their URLs.

This is created with the idea that eventually the
assembly pipeline can query ID3C to get the URLs
from which it can then grab the FASTQ files.