Add sequence read set urls unique to one set check
In PostgreSQL, arrays cannot access method GIST, which is
needed for the default exclusion constraint. It is on the
PostgreSQL [to-do-list](https://wiki.postgresql.org/wiki/Todo#Indexes)
to allow GIN indexesto be used for exclusion constraints.

For now, use triggers to check on insert and update whether
urls are unique to one sequence read set. In other words,
a specific url should only belong to one sequence read set.