reportable-conditions: Store the reporting log in a subfield of details
Instead of ambiguously at the top-level.  This will play a little nicer
with any other details we happen to have.

The reportable_conditions_v1 view will want updates to read from this
field instead of the previous two, but those columns are just
informative and aren't required for any functionality.