Update consensus genome etl
Update the consensus genome etl to create more unique identifiers
for genomic sequences by adding the sequence read set id to the end
of the identifiers.

Added an extra assert statement that verifies the consensus genome id
linked to the genomic sequence is the same as the id of the consensus
genome that had just been upserted.

I have not updated the REVISION number because we don't want the
etl process to double insert the same sequences (there will not be an
identifier conflict with previously created sequences).

The genomic sequences that are already within production will need to
be updated manually.