Locations: Use a GiST index on polygons instead of SP-GiST
The version of PostgreSQL and PostGIS we use in production doesn't
support SP-GiST indexes for geometries.  The documentation¹ indicates
that there is not a large difference between the two index types.

I'm making this tweak in place since the changes are not yet deployed
due to the lack of support.

¹ https://postgis.net/docs/using_postgis_dbmanagement.html#idm2252