location import: Switch to the older PostGIS function st_geomfromgeojson(text)
Newer versions of PostGIS also include variants of the function which
take the json and jsonb data types, but our production instance uses an
older PostGIS version.  The function which takes the text type is
supported by both older and newer versions.

This, and the previous commit, are a great example of why our
staging/testing infrastructure should be in good working order instead
of in disrepair.