location: Require EPSG:4326 (WGS84) not EPSG:4269 (NAD83)
While I got the right CRS in the schema, the location command was
enforcing the wrong one.  I clearly mixed up the numbers.

To correct existing data, a manual update can be run which does:

    st_transform(st_setsrid(polygon, 4269), 4326)

for the point, polygon, and simplified_polygon columns.

For future data which isn't already in 4326 (like official Census tract
geometries), we'll need to add support for re-projection during location
import.  Alternatively, we can re-project before import as a
pre-processing step, but I think it'd be nicer to add built-in support.