Add current and modified timestamps to warehouse
Create new columns titled 'created' and 'modified' in the following
tables:
* `warehouse.encounters`
* `warehouse.sample`
* `warehouse.presence_absence`
* `warehouse.sequence_read_set`
* `warehouse.genomic_sequence`

These columns each have default values of the current timestamp. The
`modified` column is updated with a new timestamp via a trigger function
every time a record is updated. The trigger function modifies the
`modified` column even if the update changes are identical to the
previous changes.