Tests: Rewrite dev/show-help in Python
Much faster because all commands are invoked within the same process and
also more integrated with pytest so that each command is a separate test
case.  This is nice when combined with pytest's -v option, which lists
results for individual test cases.  It also won't stop after the first
error, like dev/show-help did.