Add command `location lookup`
The `location lookup` command can take a filename option or get input
data from stdin. It will search within `warehouse.location` to find
a location of a provided scale that contains the provided lat/lng.

Currently, the lookup only returns one matching location. It
orders the query results by `location.identifier` and returns the first
location.

The location identifier is appended to the input data as a new column
named <scale>_identifier and output to stdout as comma-separated values.
Optionally, the drop-columns can be provided to drop list columns
from the output.