Add command `de-identify`
The `de-identify` command can take a filename option or get data
from stdin. It will join values from specified <columns> and hash the
joined value with a secret from the environment variable
PARTICIPANT_DEIDENTIFIER_SECRET.

The field values are joined in the order that the <columns> are
specified, so be sure to keep the order of the column names consistent!

The values of the input are expected to be canonicalized before
running the `de-identify` command.

The generated hash is appended to the input data as a new column named
"hash" and output to stdout as comma-separated values. Optionally, the
provided <columns> can be dropped using the `--drop-input-columns` flag.