etl presence-absence: Add update_presence_absence_identifier
Before the upsert of the presence/absence result, we update the
presence absence identifier from the old format ("NWGC_{test_id}")
to the new format ("NWGC/{sampleID}/{geneTarget}") if it already exists
within warehouse.presence_absence.

This is done because the {test_id} is no longer expected to be a
stable identifier.