Add a generic "receiving" command for uploading documents
Typically there is a more tailored ID3C command to manipulate specific
tables in the receiving area (such as `id3c manifest` or `id3c
redcap-det`), but sometimes it is useful to more directly manipulate
tables.  This is particularly true for receiving tables which are
generally populated through the web API and do not have a more tailored
command.