geocode: Always save the cache to disk, even on error
This ensures that if an error occurs part way through processing the
input dataset, subsequent re-runs won't have to repeat geocoding
requests made before the error.  Particularly during debugging, this
will be handy for not blowing through our geocoding quota.

A context manager is the easiest way to ensure things happen even after
error, so I wrote one for the cache loading/saving.

As a side effect, a transient, in-memory cache is now used when
--cache-file is not provided.  This simplifies our internal logic and
allows repeat addresses found in an input dataset to avoid repeat
requests even if no cache is saved.