location lookup: Optionally drop the given --lat-column and --lng-column from output
This makes it simple to take a dataset with coordinates and replace the
coordinates with an acceptable, larger scale of geography.

In Jover's original PR, there was a --drop-input-columns flag that
dropped _all_ columns in the input data.  Review notes from me and
Kairsten lead Jover, I believe, to switch to an argument list of columns
to drop instead.  My intent, which in retrospect was not clearly
communicated, was that only the given --lat-column and --lng-columns are
optionally dropped.