etl fhir: Remove age_ceiling()
This core ETL process should be agnostic to policies like age ceilings,
which can vary widely.  It is also too late from a privacy perspective
if the FHIR document in the database has an age above some threshold;
the proper time to apply a ceiling is before that FHIR document ever
hits the database (as we're doing in our custom REDCap ETLs).