etl fhir: Handle SNOMED code for "Indeterminate" in presence/absence results
This code will be sent by Ellume if their device is unable to produce a
positive/negative result.  We don't do anything with indeterminate
results in ID3C, but the Ellume device error will be in the display text
of the FHIR document in the receiving table for ad-hoc analysis if
necessary.