etl fhir: Create samples with a collection identifier for DiagnosticReports
This will safely co-operate with the manifest ETL because it is
equivalent to the way we ingest enrollments/encounters, where a sample
with a collection_identifier but no identifier is created.

The creation also happens only after verifying that the collection
identifier is valid, so we won't create samples for bogus barcodes.