logging: Decrease level for urllib3 to avoid leaking sensitive data
The SmartyStreets library we use for geocoding uses urllib3 internally,
which in turn produces DEBUG messages that can contain sensitive data in
the query string of the logged request line.  Suppress those at its
package logger.

The other debug messages urllib3 produces in practice don't seem that
useful, so adjusting the log level seems like the simplest solution.  If
we want to see the other debug messages in the future, we could
implement a custom filter to censor the query strings by rewriting the
log messages.