manifest: Genericize extraction of non-essential columns
Avoids needing to add a new, study-specific option every time we need a
new column.  This was my intent from the start, but so far it's been
slightly more pragmatic to just add new options.

Requires config file changes in our specimen-manifests repo, which I've
prepared on a corresponding branch.