etl fhir: add custom `upsert_sample` function
With the addition of the clinical data pull REDCap project, the FHIR etl
needs to be able to deal with encounter samples that only have a
sample identifier and not a collection identifier.

This custom function was copied from the manifest etl and edited to
include `encounter_id`. We may want to consolidate `upsert_sample` at
some point in the future, but this was done to ensure that nothing goes
wrong with the manifest etl.