Truly, really read values in Excel files into a DataFrame of strings
If a manifest, for example, comes in with a column of mixed types, we
don't want to error out.  We want to just treat everything as a string
and deal with problems stemming from that later instead of blocking
import.  For example, if ids lose their leading zeros because they were
accidentally stored in Excel as numbers, not text, we want to find that
out when the id fails to match during ETL ingest instead of blowing up
during parsing of a file we don't control.