Reduce log level from error → info when a sample cannot be found
This is a non-error condition for the callers, which check the return
value of find_sample() and take the appropriate action.  In some cases,
the caller will raise its own error.  In other cases, like the FHIR ETL,
the sample will be created and logging an error a) does not make sense
and b) makes the automated jobs noisy.  The decision is best left to the
caller.

The error level may have made more sense at a time when find_sample()
raised its own error instead of returning None.