etl presence-absence: Remove upgrade path from old → new result identifiers
All of our data is long-since updated, so we no longer need this code
for backwards compatibility.

Removing the code also means we no longer care about the
samples[].targetResults[].id field in the received JSON documents.  This
is useful for NWGC as they want to start omitting it (or setting it to
null) for some results in the near future.