etl fhir: Handle DiagnosticReports with sample barcodes
Previously we were only processing DiagnosticReport documents with
collection barcodes. Update the FHIR ETL to determine if a barcode is
a collection or sample barcode. If it's a sample barcode, throw an error
if the sample identifier does not exist in the database. If it's a
collection barcode, continue creating a sample for the missing barcode.
Document the reason for doing this was because Ellume sometimes sends
DiagnosticReport documents before the specimen manifest sheet is
updated.