Add the concept of a "primary" encounter location
Creates a framework for prioritizing encounter-locations by relation and
uses that in a new view that subsets warehouse.encounter_location to
just the "primary" rows.  A default set of relations and priorities is
added, but these can be tailored for each database since they are just
rows in a controlled vocabulary table.

This provides support to queries which want to only follow a single link
across the one-to-many relationship without hardcoding a specific
relation (e.g. to choose the "best" location available when the data is
sparse).