etl presence-absence: Handle new "sampleState" field
This field is being used for the results of CLIA-certified COVID-19
targets.  Other research-use-only targets continue to use
"targetResults".

Indeterminate/inconclusive results will be ingested with the "present"
column set to null, under the rationale that the presence known to be
uncertain/unknown.  These kind of results are typically because some
compound targets or target replicates are positive but some are
negative.  While the schema allows for "present is null", this is the
first time we're ingesting such data.

Several workflow states which aren't actual results are skipped during
ingestion.

Only the values "Positive", "Negative", and "Review" have been observed
so far, but the other possible values were communicated to us by NWGC.
It remains to be seen if those are accurate yet.