etl presence-absence: ingest new fields `assayName` and `assayDate`
These are new fields that Tom(NWGC) added to the JSONs we receive from
Samplify.

If `assayName` is unavailable, we still assume the presence of `chip`
indicates that the assay is "OpenArray".

Note: This does not solve our problem of old COVID-19 results being
mislabeled as "OpenArray" when they came attached to OA results.
I don't want to hard-code COVID-19 results as "TaqmanQPCR" because it
sounds like there are plans to move them to "OpenArray" in the future.