Hotfix for Pandas API errors
Two places, one likely cribbed from the other, call a non-existent
.tolist() method on the return value of .unique() on a pandas.Series.
While Series themselves have a .tolist(), the return value of .unique()
is not a Series.  (This is a bit surprising to me, yes.)

It is possible this worked before Pandas 1.0.0 and only broke later when
we upgraded, but the ExtensionArray base type that's returned from
.unique() was added in 0.23, so then again maybe not.  I don't care to
investigate further, since wrapping in list() is sufficient.

This unblocks a manifest parse job which found a duplicate and promptly
broke when trying to report it.